requirements.txt
Django==5.0
django-environ==0.11.2
psycopg2-binary==2.9.9
Pillow==10.1.0
django-ckeditor
django-cleanup==8.0.0
python-telegram-bot==20.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
gunicorn==22.0.0
drf-yasg==1.21.7
django-filter
django-jazzmin==3.0.0
django-admin-sortable2
transliterate
django-elasticsearch-dsl
elasticsearch-dsl